Refactored services
